requirements.txt
Django>=5.0,<5.2
djangorestframework>=3.15,<3.16
psycopg2-binary>=2.9
